:rocket: feat: Revisão JWT e Bcrypt

+++ api-orkut/src/app.js
@@ -4,7 +4,8 @@ const pool = require("./config/db");
 const validarUsuarios = require("./validacao/usuarios");
 const validarPost = require("./validacao/post");
 const jwt = require("jsonwebtoken");
-const auth = require("./auth/authLogin")
+const auth = require("./auth/authLogin");
+const bcrypt = require("bcrypt");
 
 
 const app = express();
@@ -15,26 +16,29 @@ app.get("/", (req, res) => {
 });
 
 // Rota Cadastro novo usuário
-// app.post("/usuarios", validarUsuarios, async (req, res) => {
-//   try {
-//     const {nome, email, senha} = req.body;
-//     const resultado = await pool.query(`
-//       INSERT INTO usuarios (nome, email, senha)
-//       VALUES ($1, $2, $3)
-//       RETURNING *
-//     `,
-//     [nome, email, senha]
-//   );
-//   res.status(201).json({
-//     mensagem: "Usuário criado com sucesso",
-//     usuario: resultado.rows[0]
-//   })
-//   } catch (erro) {
-//     res.status(500).json({
-//       erro: "Erro ao criar usuário"
-//     })
-//   }
-// })
+app.post("/usuarios", validarUsuarios, async (req, res) => {
+  try {
+    const {nome, email, senha} = req.body;
+
+    const senhaHash = await bcrypt.hash(senha, 10)
+
+    const resultado = await pool.query(`
+      INSERT INTO usuarios (nome, email, senha)
+      VALUES ($1, $2, $3)
+      RETURNING *
+    `,
+    [nome, email, senhaHash]
+  );
+  res.status(201).json({
+    mensagem: "Usuário criado com sucesso",
+    usuario: resultado.rows[0]
+  })
+  } catch (erro) {
+    res.status(500).json({
+      erro: "Erro ao criar usuário"
+    })
+  }
+})
 
 // Rota de Login
 app.post("/login", async (req, res) => {
@@ -53,7 +57,9 @@ app.post("/login", async (req, res) => {
     })
   }
 
-  if (senha !== usuario.rows[0].senha) {
+  const senhaValida = await bcrypt.compare(senha, usuario.rows[0].senha)
+
+  if (!senhaValida) {
     return res.status(400).json({
       mensagem: "Senha inválida"
     })
@@ -76,7 +82,7 @@ app.post("/login", async (req, res) => {
   }
 })
 
-
+// Rota Get usuários
 app.get("/usuarios", async (req, res) => {
   try {
     const resultado = await pool.query(`
@@ -88,6 +94,7 @@ app.get("/usuarios", async (req, res) => {
   }
 });
 
+// Rota Get posts
 app.get("/posts", async (req, res) => {
   try {
     const resultado = await pool.query(`
@@ -109,7 +116,7 @@ app.get("/posts", async (req, res) => {
   }
 });
 
-
+// Rota Post - postagem de conteúdo
 app.post("/posts", auth, validarPost, async (req, res) => {
   try {
     const { titulo, conteudo } = req.body;
@@ -132,10 +139,21 @@ app.post("/posts", auth, validarPost, async (req, res) => {
   }
 });
 
-app.put("/posts/:id", validarPost, async (req, res) => {
+// Rota Put - atualização de conteúdo
+app.put("/posts/:id", auth, validarPost, async (req, res) => {
   try {
     const { id } = req.params;
     const { titulo, conteudo } = req.body;
+
+    const post = await pool.query(`SELECT * FROM post WHERE id=$1`, [id]);
+
+    if (post.rows.length === 0) {
+      return res.status(404).json({mensagem: "Post não encontrado"})
+    };
+
+    if (post.rows[0].usuario_id !== req.usuario.id) {
+      return res.status(403).json({mensagem: "Sem permissão"})
+    };
 
     const resultado = await pool.query(
       `UPDATE post SET titulo=$1, conteudo=$2 WHERE id=$3 RETURNING *`,
@@ -152,9 +170,21 @@ app.put("/posts/:id", validarPost, async (req, res) => {
   }
 });
 
-app.delete("/posts/:id", async (req, res) => {
+//Rota Delete - apagar um conteúdo postado
+app.delete("/posts/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+
+    const post = await pool.query(`SELECT * FROM post WHERE id=$1`, [id]);
+
+    if (post.rows.length === 0) {
+      return res.status(404).json({mensagem: "Post não encontrado"})
+    };
+
+    if (post.rows[0].usuario_id !== req.usuario.id) {
+      return res.status(403).json({mensagem: "Sem permissão"})
+    }
+
     const resultado = await pool.query(
       `DELETE FROM post WHERE id=$1 RETURNING *`,
       [id],
